assets

+++ db/migrate/20170202212930_create_teams.rb
@@ -2,7 +2,7 @@ class CreateTeams < ActiveRecord::Migration[5.0]
   def change
     create_table :teams do |t|
       t.string :name
-      t.bigint :score, default: 0
+      t.decimal :score, default: 0
       t.string :color
 
       t.timestamps
